refactor: fix all clippy warnings, migrate RealHV→ContinuousHV in examples
Clippy fixes (56 warnings + 2 errors → 0):
- Fix correctness bug in foundations.rs: sum formula used 0u64 literals
  instead of parameter n, always returning zero
- Replace i64::abs() casts with unsigned_abs() for overflow safety
- Use vec![] macro instead of push-after-creation pattern
- Fix doc comment indentation across 4 files
- Add Default impls for PeanoPhysicsBridge, ConsciousnessTrajectory
- Merge identical if-else blocks in 3 files
- Fix digit grouping in physics constants
- Add targeted #[allow(clippy::...)] for intentional patterns

Example migration (14 files):
- Replace all RealHV imports with ContinuousHV across examples
- Remove redundant Vec<RealHV> conversion blocks in benchmarks
- Update comments referencing old type names

+++ symthaea/symthaea-core/src/physics/nonequilibrium.rs
@@ -334,7 +334,7 @@ impl JarzynskiEstimator {
             let mut bootstrap_sample = Vec::with_capacity(n);
             for i in 0..n {
                 // Deterministic pseudo-random selection
-                let idx = ((b * 31 + i * 17) % n) as usize;
+                let idx = (b * 31 + i * 17) % n;
                 bootstrap_sample.push(work_samples[idx]);
             }
             bootstrap_estimates.push(self.free_energy_difference(&bootstrap_sample, temperature));
